pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "mental-health-signal-detector"
version = "0.1.0"
description = ""
authors = [
    {name = "Stanislav Grinchenko",email = "stasagr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "pandas (>=1.4.0,<4.0.0)",
    "datasets (>=4.7.0,<5.0.0)",
    "kagglehub (>=1.0.0,<2.0.0)",
    "kaggle (>=2.0.0,<3.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "nltk (>=3.9.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "transformers (>=5.3.0,<6.0.0)",
    "torch (>=2.10.0,<3.0.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "streamlit (>=1.40.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "xgboost (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 150
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "D101", "D102", "D103"]  # pycodestyle, pyflakes, isort
